requirements.txt
Appium-Python-Client==0.40
mitmproxy==4.0.4
requests==2.22.0